+++ src/main/java/com/yanny/ytech/ForgeBusSubscriber.java
@@ -135,7 +135,7 @@ public class ForgeBusSubscriber {
 
             LOGGER.info("Set requiresCorrectToolForDrops on {}", Objects.requireNonNull(BuiltInRegistries.BLOCK.getKey(block)));
         } catch (Exception e) {
-            LOGGER.warn("Unable to set requiresCorrectToolForDrops on block " + Objects.requireNonNull(BuiltInRegistries.BLOCK.getKey(block)) + ": " + e.getMessage());
+            LOGGER.warn("Unable to set requiresCorrectToolForDrops on block {}: {}", Objects.requireNonNull(BuiltInRegistries.BLOCK.getKey(block)), e.getMessage());
         }
     }
 }