+++ src/main/java/com/yanny/ytech/compatibility/jei/DryingRecipeCategory.java
@@ -4,7 +4,7 @@ import com.yanny.ytech.YTechMod;
 import com.yanny.ytech.configuration.MaterialType;
 import com.yanny.ytech.configuration.Utils;
 import com.yanny.ytech.configuration.recipe.DryingRecipe;
-import com.yanny.ytech.registration.YTechBlocks;
+import com.yanny.ytech.registration.YTechItems;
 import com.yanny.ytech.registration.YTechRecipeTypes;
 import mezz.jei.api.gui.builder.IRecipeLayoutBuilder;
 import mezz.jei.api.gui.drawable.IDrawable;
@@ -38,7 +38,7 @@ public class DryingRecipeCategory implements IRecipeCategory<DryingRecipe> {
     public DryingRecipeCategory(IGuiHelper guiHelper) {
         ResourceLocation location = Utils.modLoc("textures/gui/jei.png");
         background = guiHelper.createDrawable(location, 0, 0, 82, 34);
-        icon = guiHelper.createDrawableItemStack(new ItemStack(YTechBlocks.DRYING_RACKS.of(MaterialType.OAK_WOOD).get()));
+        icon = guiHelper.createDrawableItemStack(new ItemStack(YTechItems.DRYING_RACKS.get(MaterialType.OAK_WOOD).get()));
         localizedName = Component.translatable("gui.ytech.category.drying");
     }
 
@@ -86,6 +86,6 @@ public class DryingRecipeCategory implements IRecipeCategory<DryingRecipe> {
     }
 
     public static void registerCatalyst(@NotNull IRecipeCatalystRegistration registration) {
-        YTechBlocks.DRYING_RACKS.blocks().forEach((block) -> registration.addRecipeCatalyst(new ItemStack(block.get()), RECIPE_TYPE));
+        YTechItems.DRYING_RACKS.values().forEach((item) -> registration.addRecipeCatalyst(new ItemStack(item.get()), RECIPE_TYPE));
     }
 }