+++ src/main/java/com/yanny/ytech/configuration/block/MachineBlock.java
@@ -49,7 +49,7 @@ public abstract class MachineBlock extends BaseEntityBlock {
 
     @Override
     protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> stateBuilder) {
-        stateBuilder.add(BlockStateProperties.HORIZONTAL_FACING, BlockStateProperties.POWERED);
+        stateBuilder.add(BlockStateProperties.HORIZONTAL_FACING, BlockStateProperties.LIT);
     }
 
     @Override
@@ -63,7 +63,7 @@ public abstract class MachineBlock extends BaseEntityBlock {
     public BlockState getStateForPlacement(@NotNull BlockPlaceContext blockPlaceContext) {
         return defaultBlockState()
                 .setValue(BlockStateProperties.HORIZONTAL_FACING, blockPlaceContext.getHorizontalDirection().getOpposite())
-                .setValue(BlockStateProperties.POWERED, false);
+                .setValue(BlockStateProperties.LIT, false);
     }
 
     @NotNull