+++ src/main/java/com/yanny/ytech/configuration/recipe/RemainingShapedRecipe.java
@@ -48,10 +48,7 @@ public class RemainingShapedRecipe extends ShapedRecipe {
                 MinecraftServer server = ServerLifecycleHooks.getCurrentServer();
 
                 if (server != null) {
-                    result.hurtAndBreak(1, server.overworld(), null, (it) -> {
-                        result.shrink(1);
-                        result.setDamageValue(0);
-                    });
+                    result.hurtAndBreak(1, server.overworld(), null, (it) -> {});
                 }
             }
         }